move data visualization content
