+++ src/app/core-components/sidebar/sidebar.component.html
@@ -17,7 +17,7 @@
         <a class="nav-link" routerLink="/account/reports" [routerLinkActive]="['active']"><i class="fa fa-folder"></i>
         </a>
         <a class="nav-link disabled" routerLink="/account/analysis" [routerLinkActive]="['active']"><i
-                class="fa fa-area-chart"></i>
+                class="fa fa-microscope"></i>
         </a>
         <a class="nav-link" routerLink="/account/settings" [routerLinkActive]="['active']"><i class="fa fa-cog"></i>
         </a>
@@ -38,8 +38,11 @@
             <a class="nav-link" routerLink="/facility/{{facility.id}}/utility" [routerLinkActive]="['active']"><i
                     class="fa fa-file-invoice"></i>
             </a>
+            <a class="nav-link" routerLink="/facility/{{facility.id}}/visualization" [routerLinkActive]="['active']"><i
+                    class="fa fa-chart-line"></i>
+            </a>
             <a class="nav-link" routerLink="/facility/{{facility.id}}/analysis" [routerLinkActive]="['active']"><i
-                    class="fa fa-area-chart"></i>
+                    class="fa fa-microscope"></i>
             </a>
             <a class="nav-link" routerLink="/facility/{{facility.id}}/settings" [routerLinkActive]="['active']"><i
                     class="fa fa-cog"></i>
