+++ src/app/facility/facility.component.html
@@ -2,4 +2,3 @@
 <div class="dashboard-container facility">
     <router-outlet></router-outlet>
 </div>
-<div class="measurFooter"></div>
