+++ src/app/facility/utility-data/utility-banner/utility-banner.component.html
@@ -14,9 +14,6 @@
         <li class="nav-item" [class.disabled]="utilityMeterData.length === 0">
             <a class="nav-link" routerLink="predictors" [routerLinkActive]="['active']">Predictors</a>
         </li>
-        <li class="nav-item" [class.disabled]="utilityMeterData.length === 0">
-            <a class="nav-link" routerLink="visualization" [routerLinkActive]="['active']">Data Visualization</a>
-        </li>
         <li class="nav-item">
             <a class="nav-link" routerLink="upload-data" [routerLinkActive]="['active']">Import Data</a>
         </li>
