+++ src/app/facility/utility-data/utility-data.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { UtilityDataComponent } from './utility-data.component';
-import { VisualizationModule } from './visualization/visualization.module';
+import { VisualizationModule } from '../visualization/visualization.module';
 import { UploadDataModule } from './upload-data/upload-data.module';
 import { PredictorDataModule } from './predictor-data/predictor-data.module';
 import { MeterGroupingModule } from './meter-grouping/meter-grouping.module';
